ci(failure) Fail build when classpath projects can't be generated
- Fix missing arguments for older-versions
- Manually handle exit codes since Travis doesn't respect exit codes at all !!!

+++ ci/generate-classpath-tests.sh
@@ -9,18 +9,24 @@ GENERATION_ARGS=$([ "$profile" = "" ] && echo ''  || echo '-Dspring.profiles.act
 
 # start Spring Initializr
 touch "$HOME/$INITIALIZER_NAME.log"
-spring-initializr/./mvnw clean spring-boot:run -f spring-initializr > "$HOME/$INITIALIZER_NAME.log" &
+(spring-initializr/./mvnw clean spring-boot:run -f spring-initializr $INITIALIZER_ARGS > "$HOME/$INITIALIZER_NAME.log") &
 echo "$!" > "$HOME/$INITIALIZER_NAME.pid"
-# wait until started
-(tail -f "$HOME/$INITIALIZER_NAME.log" &) | grep -q 'Started SpringInitializrApplication'
+# wait until started or failed
+(tail -f "$HOME/$INITIALIZER_NAME.log" &) | grep -q -e 'Started SpringInitializrApplication' -e 'BUILD FAILURE'
+# if failed => log and stop
+if grep 'BUILD FAILURE' "$HOME/$INITIALIZER_NAME.log"
+then
+	echo "Failed to start spring-initializr:"
+	cat "$HOME/$INITIALIZER_NAME.log"
+	sleep 2s
+	exit 1
+fi
+
 
 # Generate projects
 OGHAM_VERSION=$(./mvnw -q -Dexec.executable="echo" -Dexec.args='${project.version}' --non-recursive org.codehaus.mojo:exec-maven-plugin:1.3.1:exec)
 ./mvnw spring-boot:run -pl :ogham-test-classpath -Dspring-boot.run.arguments="$HOME/classpath-tests" -Dspring-boot.run.jvmArguments="-Drunner.parallel=false -Dogham-version=$OGHAM_VERSION -Dspring.initializer.url=http://localhost:$INITIALIZER_PORT/starter.zip $GENERATION_ARGS"
 
-ls -l "$HOME"
-ls -l "$HOME/classpath-tests"
 ls -l "$HOME/classpath-tests/$TEST_FOLDER"
 
 kill `cat "$HOME/$INITIALIZER_NAME.pid"` || echo ""
-
